+++ CHANGELOG.md
@@ -1,5 +1,99 @@
 # Changelog
 
+## [2.6.1](https://github.com/Viren070/AIOStreams/compare/v2.6.0...v2.6.1) (2025-07-16)
+
+
+### Bug Fixes
+
+* correctly extract extras for subtitle requests ([a1cbeb2](https://github.com/Viren070/AIOStreams/commit/a1cbeb2880a4abaf43f3c2fe9eabab8239e024df))
+* **debridio-tv:** add new zealand option ([3dd2e7a](https://github.com/Viren070/AIOStreams/commit/3dd2e7a3c5189da1b31dbc6f3044477cd065dc91))
+* **fkstream:** allow overriding services ([952dda8](https://github.com/Viren070/AIOStreams/commit/952dda8bd106a5176d12d554280e620219576d12)), closes [#268](https://github.com/Viren070/AIOStreams/issues/268)
+
+## [2.6.0](https://github.com/Viren070/AIOStreams/compare/v2.5.4...v2.6.0) (2025-07-16)
+
+
+### Features
+
+* add ai search addon to marketplace ([c9c1bab](https://github.com/Viren070/AIOStreams/commit/c9c1babc2acaa5d92e3230382bf3dcfca2f874d5))
+* add AIO Subtitle addon to marketplace ([db191fd](https://github.com/Viren070/AIOStreams/commit/db191fd430b17b8452fc6e4a0b2f776f9463f3b3))
+* add fkstream to marketplace ([ee72f3a](https://github.com/Viren070/AIOStreams/commit/ee72f3a1ed4f137ba07a1e23c381b257acaa2bde)), closes [#260](https://github.com/Viren070/AIOStreams/issues/260)
+* add statistics position option to control where statistic streams appear ([47656d0](https://github.com/Viren070/AIOStreams/commit/47656d0413996ae475b6c25d9503fb60595f8429))
+* add subhero to the marketplace ([e70920a](https://github.com/Viren070/AIOStreams/commit/e70920a2701db19796146c486558a59939e616cd))
+* extract episode count during precaching to automatically precache next season when necessary ([6f0e79a](https://github.com/Viren070/AIOStreams/commit/6f0e79a5d9a8951e882251dc7a48d7a70046a0d8))
+* **frontend:** move tmdb access token setting to services menu ([9101acb](https://github.com/Viren070/AIOStreams/commit/9101acb72f02ec5d5573e41e138797040ef39c0a))
+* **webstreamr:** update provider list and automatically provide mediaflow details if possible ([dbde122](https://github.com/Viren070/AIOStreams/commit/dbde1225e4b7c3505009b97e1454fd3f09840625))
+
+
+### Bug Fixes
+
+* add fkstream, subhero, aio subtitle to startup logs and sample .env ([130be7d](https://github.com/Viren070/AIOStreams/commit/130be7dc16ce2e4e37d59d7ff3c66afe5b08f5b6))
+* add socials option to ai search ([28f7863](https://github.com/Viren070/AIOStreams/commit/28f786339e12b182856764d31605c76610a75335))
+* adjust extras schema to allow more than 1 extra ([7ba880e](https://github.com/Viren070/AIOStreams/commit/7ba880e9a74b5f6e7b49b2bcff8dbda6b13a159d)), closes [#263](https://github.com/Viren070/AIOStreams/issues/263)
+* **aiosubtitle:** rename language option to languages and use correct default timeout ([603af9a](https://github.com/Viren070/AIOStreams/commit/603af9a7e97226272300377f5ac6dc2bb62e816b))
+* allow any string for thumbnail ([4652eb6](https://github.com/Viren070/AIOStreams/commit/4652eb650273f780cb77e42c8aaab34e63ce90a0))
+* **core:** set groups to undefined during precaching ([02d4cf7](https://github.com/Viren070/AIOStreams/commit/02d4cf742b0d6be75b4cf281b47ef7dd75a9affb))
+* **core:** use structuredClone for cache item values to ensure immutability ([465b7a4](https://github.com/Viren070/AIOStreams/commit/465b7a40a34909bc84290da37cefbe82f3131ea4))
+* correct spelling of ALLEDEBRID_SERVICE to ALLDEBRID_SERVICE ([b67f38a](https://github.com/Viren070/AIOStreams/commit/b67f38ac85033f48efea412f90945f7f15802ee0))
+* **debridio-tv:** update channel list ([67509a4](https://github.com/Viren070/AIOStreams/commit/67509a43def9e01250beaa62bddb6207d399b5be))
+* ensure number types always have integer step of at least 1 ([ac4d209](https://github.com/Viren070/AIOStreams/commit/ac4d209879b64a64d7df600a2b01c8164fe726c5))
+* ensure string validations are carried out against password type ([66d4a96](https://github.com/Viren070/AIOStreams/commit/66d4a96ad20cea105cc23c19945febdf4cf10e56))
+* **frontend:** revert 3a82b796c56627d21ede53b91f47315cbb46c1e4 and clean options before processing ([880817a](https://github.com/Viren070/AIOStreams/commit/880817a1ad100799826dbde10e87c303b74b958e))
+* **frontend:** show puzzle icon for addons without logo ([f74fbb2](https://github.com/Viren070/AIOStreams/commit/f74fbb289b208f1ea082e74675b7bd274153d5f4))
+* **frontend:** strictly rely on value during selection ([3a82b79](https://github.com/Viren070/AIOStreams/commit/3a82b796c56627d21ede53b91f47315cbb46c1e4))
+* improve parsing for debridio watchtower, nuviostreams, and webstreamr ([9febeeb](https://github.com/Viren070/AIOStreams/commit/9febeeb86af023fed32dc25dc89dfdad09406886))
+* **nuviostreams:** add animepahe provider ([00862c0](https://github.com/Viren070/AIOStreams/commit/00862c05926a18766592bc3a177075b14ced90a4))
+* only move to next season when it exists and correct logs ([3050b41](https://github.com/Viren070/AIOStreams/commit/3050b41e937fa24dcfb146a953862c4054cf4a0f))
+* remove unnecessary debug logs ([0e24a6a](https://github.com/Viren070/AIOStreams/commit/0e24a6aa961c0dee200ea018caec06b23134d6cb))
+* replace discover deep links where possible ([aa084d3](https://github.com/Viren070/AIOStreams/commit/aa084d35d646f05cc2dea7111ad34ba5d1a6d661)), closes [#264](https://github.com/Viren070/AIOStreams/issues/264)
+
+
+### Performance Improvements
+
+* always fetch from all groups in parallel ([0ce1338](https://github.com/Viren070/AIOStreams/commit/0ce13389a066a05184bc1aaea0b2806b22d01a1f))
+
+## [2.5.4](https://github.com/Viren070/AIOStreams/compare/v2.5.3...v2.5.4) (2025-07-12)
+
+
+### Bug Fixes
+
+* ensure catalog.extra is initialised before pushing new genre options ([dc9c8a4](https://github.com/Viren070/AIOStreams/commit/dc9c8a40775ba13d66ff7a762147b50e4b414974))
+
+## [2.5.3](https://github.com/Viren070/AIOStreams/compare/v2.5.2...v2.5.3) (2025-07-11)
+
+
+### Bug Fixes
+
+* only add aiostreamserror to idPrefixes if already defined ([ebe46d7](https://github.com/Viren070/AIOStreams/commit/ebe46d7c43acc70fdb52f7aeca41aeda27bf7df7))
+* only log warning for missing idPrefixes when needed ([08ec779](https://github.com/Viren070/AIOStreams/commit/08ec779be3bd307e0fda9766301d80fdc39a8f3e))
+
+## [2.5.2](https://github.com/Viren070/AIOStreams/compare/v2.5.1...v2.5.2) (2025-07-11)
+
+
+### Features
+
+* add statistic stream options to see statistics in stremio ([261b878](https://github.com/Viren070/AIOStreams/commit/261b87872a2882886ec26970614a504dc2f7cd97))
+
+
+### Bug Fixes
+
+* account for query parameters in manifest urls for custom addons during validation ([3ad1e7c](https://github.com/Viren070/AIOStreams/commit/3ad1e7c61f36d12dfb046c676a401b0e3f97604b))
+* add skipReasons logging for seeder ranges ([a06ac34](https://github.com/Viren070/AIOStreams/commit/a06ac34061e571fe7599eeca5a0fce6d9607b37f))
+* correct spelling of 'Crunchyroll' in streaming catalogs preset ([cb423d1](https://github.com/Viren070/AIOStreams/commit/cb423d152dc015b37b3d317c80907b357f7ee154))
+* **debridio-watchtower:** ensure resolution is always parsed ([6a0bf09](https://github.com/Viren070/AIOStreams/commit/6a0bf09de660198cd317bcd97ec8777aa4723e74))
+* **debridio-watchtower:** update stream parser ([542863e](https://github.com/Viren070/AIOStreams/commit/542863e2a807d18cd6d32c7ba8fa111a711ec37a))
+* forward manifest parsing errors ([9748ecb](https://github.com/Viren070/AIOStreams/commit/9748ecb576320db85c3a5c81a9d67e8f80dd5a4c))
+* **frontend:** remove menu query param on start/'about' menu ([275a2de](https://github.com/Viren070/AIOStreams/commit/275a2de9605c01911714574a040d27f2c3977c76))
+* only log warning for missing idPrefixes for non 'catlaog' resources ([f8065c9](https://github.com/Viren070/AIOStreams/commit/f8065c9e9ad6959c51f664f377ee75a221a92c32))
+* only standardise upon valdiation ([e980c1e](https://github.com/Viren070/AIOStreams/commit/e980c1e0b3a5026b5efdfd8bd812b6bfef90bc45))
+* pass fileIdx through for p2p streams ([0076339](https://github.com/Viren070/AIOStreams/commit/00763397a1b2952f2a99c2e57e8c275863d5aa54))
+* provide meta for catalog errors ([f4050d6](https://github.com/Viren070/AIOStreams/commit/f4050d6f388a774751412f89a1e1ff1210aa76c0))
+* set customised logo to undefined when empty ([7d0b60e](https://github.com/Viren070/AIOStreams/commit/7d0b60efa1f4ec1fc2de6402ffe8b2f6b2c9d2ca))
+* set loading to false in the case of an error upon fetching the new manifest URL ([707a1f3](https://github.com/Viren070/AIOStreams/commit/707a1f3fecb781650078d5c3ca7db3842c102869))
+* standardise manifest URL during validation for custom addon updates ([8314cae](https://github.com/Viren070/AIOStreams/commit/8314cae4811d7e2307498dc0cda12b7822518708))
+* update director to allow string ([2c924ba](https://github.com/Viren070/AIOStreams/commit/2c924bac56e2a3d81cc1fe6ae380286e7fe93b58))
+* update publicIp schema to allow empty string in addition to valid IP ([0539fe4](https://github.com/Viren070/AIOStreams/commit/0539fe49a8cfe8d18525262256388a60691b8e23)), closes [#250](https://github.com/Viren070/AIOStreams/issues/250)
+* use correct property in skipReasons for seeder ranges ([c54b2b6](https://github.com/Viren070/AIOStreams/commit/c54b2b66f050827b6e41db1bf486da3140683487))
+
 ## [2.5.1](https://github.com/Viren070/AIOStreams/compare/v2.5.0...v2.5.1) (2025-07-04)